+++ mirar/database/errors.py
@@ -1,0 +1,8 @@
+"""
+Database errors
+"""
+from mirar.errors.exceptions import ProcessorError
+
+
+class DataBaseError(ProcessorError):
+    """Error relating to postgres interactions"""
