+++ mirar/database/transactions/__init__.py
@@ -1,0 +1,4 @@
+"""
+Central module for all DB transaction types.
+"""
+from mirar.database.transactions.export import export_to_db
