+++ mirar/pipelines/wirc/wirc_files/__init__.py
@@ -3,8 +3,6 @@ Module containing WIRC-specific paths
 """
 from pathlib import Path
 
-from mirar.processors.database.utils import get_column_names_from_schema
-
 wirc_file_dir = Path(__file__).parent.joinpath("files")
 wirc_schema_dir = Path(__file__).parent.joinpath("schema")
 
@@ -30,7 +28,7 @@ swarp_sp_path = wirc_file_dir.joinpath("second_pass.swarp")
 
 wirc_candidate_schema_path = wirc_schema_dir.joinpath("candidates.sql")
 
-candidate_colnames = get_column_names_from_schema(wirc_candidate_schema_path)
+# candidate_colnames = get_column_names_from_schema(wirc_candidate_schema_path)
 
 sextractor_reference_config = {
     "config_path": wirc_file_dir.joinpath("photomCat.sex"),
