+++ mirar/processors/database/base_database_processor.py
@@ -3,10 +3,11 @@ Module containing base database processor class
 """
 import logging
 from abc import ABC
+from typing import Type
 
+from mirar.database.base_model import BaseDB
 from mirar.database.constants import POSTGRES_DUPLICATE_PROTOCOLS
-from mirar.database.user import PostgresUser
-from mirar.paths import max_n_cpu
+from mirar.database.user import PostgresAdmin, PostgresUser
 from mirar.processors.base_processor import BaseProcessor
 
 logger = logging.getLogger(__name__)
@@ -15,24 +16,24 @@ logger = logging.getLogger(__name__)
 class BaseDatabaseProcessor(BaseProcessor, ABC):
     """Base class for processors which interact with a postgres database"""
 
-    max_n_cpu = min(max_n_cpu, 5)
+    max_n_cpu = 1
 
     def __init__(
         self,
-        db_name: str,
-        db_table: str,
+        db_table: Type[BaseDB],
         pg_user: PostgresUser = PostgresUser(),
+        pg_admin: PostgresAdmin = PostgresAdmin(),
         duplicate_protocol: str = "fail",
         q3c_bool: bool = False,
     ):
         super().__init__()
-        self.db_name = db_name
         self.db_table = db_table
+        self.db_name = self.db_table.sql_model.db_name
+        self.db_check_bool = False
+        self.duplicate_protocol = duplicate_protocol
+        self.q3c = q3c_bool
 
         self.pg_user = pg_user
-
-        self.duplicate_protocol = duplicate_protocol
+        self._pg_admin = pg_admin
 
         assert self.duplicate_protocol in POSTGRES_DUPLICATE_PROTOCOLS
-
-        self.q3c = q3c_bool
