--- mirar/processors/sqldatabase/__init__.py
@@ -1,4 +1,0 @@
-"""
-Central module for interactions with postgres databases
-"""
-from mirar.processors.sqldatabase.base_database_processor import BaseDatabaseProcessor
